requirements.txt
# Alpha Hunter Signals - Requirements
# Core dependencies for autonomous options trading signals

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Market Data APIs
yfinance>=0.2.18
alpha_vantage>=2.3.1
finnhub-python>=2.4.18
polygon-api-client>=1.9.1

# Web Scraping & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.8.0

# Machine Learning
scikit-learn>=1.3.0
pandas-ta>=0.3.14b0

# Telegram Integration
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Data Validation & Processing
pydantic>=2.0.0
jsonschema>=4.19.0

# Trading & Financial
alpaca-trade-api>=3.0.0
py-vollib>=1.0.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
cachetools>=5.3.0

# Optional: Enhanced Features
# openai>=1.0.0  # For enhanced AI analysis
# anthropic>=0.3.0  # For Claude integration